Updated debug output using levels of output Clean up version
